Seq2Seq-汉译英

+++ translation-seq2seq/src/dataset.py
@@ -14,7 +14,7 @@ class TranslationDataset(Dataset):
 
     def __getitem__(self, index):
         input_tensor = torch.tensor(self.data[index]['zh'], dtype=torch.long)
-        target_tensor = torch.tensor(self.data[index]['en'], dtype=torch.float32)
+        target_tensor = torch.tensor(self.data[index]['en'], dtype=torch.long)
         return input_tensor, target_tensor
 
 
@@ -24,13 +24,14 @@ def get_dataloader(train=True):
     dataset = TranslationDataset(data_path)
     return DataLoader(dataset, batch_size=config.BATCH_SIZE, shuffle=True)
 
+
 if __name__ == '__main__':
     train_dataloader = get_dataloader(train=True)
     print(f'train batch个数：{len(train_dataloader)}')
     test_dataloader = get_dataloader(train=False)
     print(f'test batch个数：{len(test_dataloader)}')
 
-    for inputs,targets in train_dataloader:
-        print(f'inputs.shape:{inputs.shape}') # [batch_size, seq_len]
-        print(f'targets.shape:{targets.shape}') # [batch_size]
+    for inputs, targets in train_dataloader:
+        print(f'inputs.shape:{inputs.shape}')  # [batch_size, seq_len]
+        print(f'targets.shape:{targets.shape}')  # [batch_size,seq_len]
         break
